make pretty log requirements optional
