+++ TODO
@@ -1,4 +1,4 @@
 ✔ pretty logger @done
-☐ command to install pretty logger requirements
-☐ pretty logger docs
+✔ command to install pretty logger requirements @done
+✔ pretty logger docs @done
 ☐ tests
